Greeting and Shortcuts are now both options
It is now possible to turn both greeting and shortcuts feature on/off
via options.

A table has also been added to the options page, though it isn't used
for anything yet.  The idea is that this table is turned into an
editable table in which users specify details about their shortcuts.